% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_cas.R
\name{cas_adjustment}
\alias{cas_adjustment}
\title{Create cluster adjusted dataset}
\usage{
cas_adjustment(
  d,
  model = NULL,
  adjust = c("all", "intercept", "none"),
  pred_fixed = NULL,
  pred_cond = NULL,
  pred_int = NULL,
  outcome_name = NULL
)
}
\arguments{
\item{d}{The dataset with clustered data}

\item{model}{A model. Defaults to NULL.}

\item{adjust}{The type of cluster adjustment (all, intercept, or none)}

\item{pred_fixed}{Prediction of the fixed effect. Defaults to NULL}

\item{pred_cond}{Prediction of the conditional effect. Defaults to NULL.}

\item{pred_int}{Prediction of the interceopt. Defaults to NULL.}

\item{outcome_name}{The name of the DV. Defaults to NULL. 

There are a lot of NULL options to make it easier for testing. The best approach
is to fit a model first, then provide that as input to the model. If not, you have
to fit the model outside of the function, which is pretty lame. See examples}
}
\value{
a dataset with the dependency in the data removed
}
\description{
Create cluster adjusted dataset
}
\examples{
mod1 = lme4::lmer(y~x + (x | id), data=small_mixed)
new_small  = cas_adjustment(small_mixed, model = mod1, adjust = "all")
}
